refactor: `universal_tools` -> `common/tools`
