+++ src/extension/tools/index.ts
@@ -10,7 +10,7 @@ import { TabManagement } from './tab_management';
 import { WebSearch } from './web_search';
 import { RequestLogin } from './request_login';
 
-export * from '../../universal_tools';
+export * from '../../common/tools';
 
 export {
   BrowserUse,
