+++ src/nodejs/tools/index.ts
@@ -1,4 +1,4 @@
-export * from '../../universal_tools';
+export * from '../../common/tools';
 
 export { CommandExecute } from './command_execute';
 export { FileRead } from './file_read';
